+++ src/api/raw_models.ts
@@ -1,5 +1,19 @@
 import type { ID, ProjectID, SectionID, LabelID } from "./models";
 
+export const UnknownProject: IProjectRaw = {
+  id: -1,
+  parent_id: null,
+  order: -1,
+  name: "Unknown project",
+};
+
+export const UnknownSection: ISectionRaw = {
+  id: -1,
+  project_id: -1,
+  order: -1,
+  name: "Unknown section",
+};
+
 export interface ITaskRaw {
   id: ID;
   project_id: ProjectID;
